@@ -5,55 +5,56 @@
     "scripts": {
         "ng": "ng",
         "start": "ng serve",
-        "build": "ng build --prod",
+        "build": "ng build",
         "test": "ng test",
         "lint": "ng lint",
         "e2e": "ng e2e"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "7.1.2",
-        "@angular/cdk": "7.1.1",
-        "@angular/common": "7.1.2",
-        "@angular/compiler": "7.1.2",
-        "@angular/core": "7.1.2",
-        "@angular/forms": "7.1.2",
-        "@angular/http": "7.1.2",
-        "@angular/platform-browser": "7.1.2",
-        "@angular/platform-browser-dynamic": "7.1.2",
-        "@angular/router": "7.1.2",
+        "@angular/animations": "8.0.0",
+        "@angular/cdk": "8.0.0",
+        "@angular/common": "8.0.0",
+        "@angular/compiler": "8.0.0",
+        "@angular/core": "8.0.0",
+        "@angular/forms": "8.0.0",
+        "@angular/platform-browser": "8.0.0",
+        "@angular/platform-browser-dynamic": "8.0.0",
+        "@angular/router": "8.0.0",
+        "@fullcalendar/core": "4.0.2",
+        "@fullcalendar/daygrid": "4.0.1",
+        "@fullcalendar/interaction": "4.0.2",
+        "@fullcalendar/timegrid": "4.0.1",
         "chart.js": "2.7.3",
-        "core-js": "2.5.7",
-        "fullcalendar": "4.0.0-alpha.2",
         "intl": "1.2.5",
-        "primeng": "7.0.0",
+        "primeng": "8.0.0",
         "quill": "1.1.8",
-        "rxjs": "6.3.3",
-        "zone.js": "0.8.26",
-        "web-animations-js": "2.3.1",
+        "rxjs": "6.4.0",
+        "web-animations-js": "github:angular/web-animations-js#release_pr208",
+        "zone.js": "0.9.1",
         "font-awesome": "4.7.0",
         "primeicons": "1.0.0",
         "prismjs": "1.15.0"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "0.11.2",
-        "@angular/compiler-cli": "7.1.2",
-        "@angular/cli": "7.1.2",
-        "@angular/language-service": "7.1.2",
+        "@angular-devkit/build-angular": "0.800.2",
+        "@angular/compiler-cli": "8.0.0",
+        "@angular/cli": "8.0.0",
+        "@angular/language-service": "8.0.0",
         "@types/node": "8.9.5",
-        "@types/jasmine": "2.8.12",
-        "@types/jasminewd2": "2.0.6",
-        "codelyzer": "4.5.0",
-        "jasmine-core": "2.99.1",
+        "@types/jasmine": "3.3.13",
+        "@types/jasminewd2": "2.0.3",
+        "codelyzer": "5.1.0",
+        "jasmine-core": "3.4.0",
         "jasmine-spec-reporter": "4.2.1",
-        "karma": "3.1.3",
+        "karma": "4.1.0",
         "karma-chrome-launcher": "2.2.0",
-        "karma-coverage-istanbul-reporter": "2.0.4",
-        "karma-jasmine": "1.1.2",
-        "karma-jasmine-html-reporter": "0.2.2",
-        "protractor": "5.4.1",
+        "karma-coverage-istanbul-reporter": "2.0.5",
+        "karma-jasmine": "2.0.1",
+        "karma-jasmine-html-reporter": "1.4.2",
+        "protractor": "5.4.2",
         "ts-node": "7.0.1",
-        "tslint": "5.11.0",
-        "typescript": "3.1.6"
+        "tslint": "5.17.0",
+        "typescript": "3.4.5"
     }
 }
