@@ -15,7 +15,7 @@ export class AppMenuComponent implements OnInit, AfterViewInit {
 
     model: any[];
 
-    @ViewChild('layoutMenuScroller') layoutMenuScrollerViewChild: ScrollPanel;
+    @ViewChild('layoutMenuScroller', { static: false }) layoutMenuScrollerViewChild: ScrollPanel;
 
     constructor(public app: AppComponent) { }
 
@@ -326,10 +326,10 @@ export class AppMenuComponent implements OnInit, AfterViewInit {
     }
 
     changeTheme(theme: string, scheme: string) {
-        const layoutLink: HTMLLinkElement = <HTMLLinkElement>document.getElementById('layout-css');
+        const layoutLink: HTMLLinkElement = document.getElementById('layout-css') as HTMLLinkElement;
         layoutLink.href = 'assets/layout/css/layout-' + theme + '.css';
 
-        const themeLink: HTMLLinkElement = <HTMLLinkElement>document.getElementById('theme-css');
+        const themeLink: HTMLLinkElement = document.getElementById('theme-css') as HTMLLinkElement;
         themeLink.href = 'assets/theme/' + theme + '/theme-' + scheme + '.css';
     }
 
@@ -430,7 +430,7 @@ export class AppSubMenuComponent {
 
         // execute command
         if (item.command) {
-            item.command({ originalEvent: event, item: item });
+            item.command({ originalEvent: event, item });
         }
 
         // prevent hash change
