@@ -1,7 +1,11 @@
 import { browser, by, element } from 'protractor';
 
 export class EcuadorPage {
-  navigateTo() {
-    return browser.get('/');
-  }
+    navigateTo() {
+        return browser.get(browser.baseUrl) as Promise<any>;
+    }
+
+    getTitleText() {
+        return element(by.css('app-root h1')).getText() as Promise<string>;
+    }
 }
