@@ -9,7 +9,7 @@ export class EventService {
     getEvents() {
         return this.http.get<any>('assets/demo/data/scheduleevents.json')
                     .toPromise()
-                    .then(res => <any[]> res.data)
+                    .then(res => res.data as any[])
                     .then(data => data);
     }
 }
