@@ -5,6 +5,9 @@ import { Car } from '../domain/car';
 import { SelectItem } from 'primeng/primeng';
 import { MenuItem } from 'primeng/primeng';
 import { BreadcrumbService } from '../../breadcrumb.service';
+import dayGridPlugin from '@fullcalendar/daygrid';
+import timeGridPlugin from '@fullcalendar/timegrid';
+import interactionPlugin from '@fullcalendar/interaction';
 
 @Component({
     templateUrl: './dashboard.component.html'
@@ -80,7 +83,12 @@ export class DashboardDemoComponent implements OnInit {
         ];
 
         this.fullcalendarOptions = {
-            defaultDate: '2016-01-12'
+            plugins: [ dayGridPlugin, timeGridPlugin, interactionPlugin ],
+            defaultDate: '2016-01-12',
+            header: {
+                right: 'prev,next, today',
+                left: 'title'
+            }
         };
     }
 }
